utils/file.go
package utils

import (
	"encoding/json"
	"fmt"
	"io/ioutil"
	"os"

	"github.com/wakahkah/go-resume-generator/model"
)

func GetDataFromFile(filePath string) (*model.ResumeData, error) {
	// Open our jsonFile
	jsonFile, err := os.Open(filePath)
	// if we os.Open returns an error then handle it
	if err != nil {
		return nil, err
	}

	fmt.Printf("Successfully Opened %s \n.", filePath)
	// defer the closing of our jsonFile so that we can parse it later on
	defer jsonFile.Close()

	// read our opened xmlFile as a byte array.
	byteValue, _ := ioutil.ReadAll(jsonFile)

	// we initialize our Users array
	var data model.ResumeData

	// we unmarshal our byteArray which contains our
	// jsonFile's content into 'users' which we defined above
	json.Unmarshal(byteValue, &data)

	return &data, nil
}
